Created a service generator
So there is no need to duplicate code each time you do a call to and endpoint.It's injected in the datasource.
